YTDB-558: tests cleanup
